refactor and modularize csrf handling, rename api handler, fix token retrieval
