--- src/api_calls/ChatContext.jsx
@@ -1,73 +1,0 @@
-const VITE_API_BASE_URL = import.meta.env.PROD
-    ? 'https://oh-kale-no-backend.onrender.com'
-    : 'http://localhost:8000';
-
-
-async function fetchCsrfToken() {
-    try {
-        console.log('Fetching CSRF token...');
-        const response = await fetch(`${VITE_API_BASE_URL}/core/get-csrf-token/`, {
-            method: 'GET',
-            credentials: 'include',
-            headers: {
-                'Accept': 'application/json',
-                'X-Requested-With': 'XMLHttpRequest',
-            }
-        });
-        
-        if (!response.ok) {
-            throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-        
-        await new Promise(resolve => setTimeout(resolve, 100));
-
-        // Get token from cookie
-        const cookieToken = document.cookie
-            .split('; ')
-            .find(row => row.startsWith('csrftoken='))
-            ?.split('=')[1];
-
-        console.log('cookieToken returned from fetchCsrfToken:', cookieToken);
-        console.log('document.cookie in fetchCsrfToken: ', document.cookie);
-        if (!cookieToken) {
-            throw new Error('CSRF token not found in cookies.');
-        }
-        
-        return cookieToken; 
-    } catch (error) {
-        console.error('Error fetching CSRF token: ', error);
-        throw error;
-    }
-}
-
-export async function getClaudeResponse(userPrompt) {
-    const csrftoken = await fetchCsrfToken();
-    try {
-        console.log('Using CSRF token for Claude request: ', csrftoken);
-        console.log('Current cookies:', document.cookie);
-
-        const response = await fetch(`${VITE_API_BASE_URL}/core/get-claude-response/`, {
-            method: 'POST',
-            credentials: 'include',
-            headers: {
-                'Content-Type': 'application/json',
-                'X-CSRFToken': csrftoken,  
-                'X-Requested-With': 'XMLHttpRequest',
-            },
-            body: JSON.stringify({ userPrompt })
-        });
-
-        if (!response.ok) {
-            const errorText = await response.text();
-            console.error('Response not OK: ', response.status, errorText);
-            console.error('Request headers: ', Object.fromEntries([...response.headers]));
-            throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-
-        const data = await response.json();
-        return data.message;
-    } catch (error) {
-        console.error('Error fetching Claude response: ', error);
-        return "Sorry, there seems to be a veggie jam in my Ninja. Please try again later.";
-    }
-}
