+++ vite.config.js
@@ -4,4 +4,13 @@ import react from '@vitejs/plugin-react'
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  server: {
+    proxy: {
+      '/core': {
+        target: 'http://localhost:8000',
+        changeOrigin: true,
+        secure: false,
+        credentials: 'include',
+      }
+    }
 })
